Added Loader and refactor
