+++ src/App/App.css
@@ -57,3 +57,9 @@ img.cover {
   font-size: 30px;
   margin: auto;
 }
+
+a.album-play {
+  font-size: 50px;
+  color: white;
+  margin-left: -10px;
+}
