+++ src/App/App.js
@@ -1,13 +1,17 @@
-import React, { Component } from 'react';
+import React, { lazy, Component } from 'react';
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router, Route, NavLink } from "react-router-dom";
 
-import Login from './Routes/Login';
-import Albums from './Routes/Albums';
-import Home from './Routes/Home';
+import Login from './Pages/Login';
+//import Albums from './Routes/Albums';
+import Home from './Pages/Home';
+import Loader from './Services/Loader';
 
 // Css
 import './App.css';
+
+
+const Albums = lazy(() => import('./Pages/Albums'));
 
 class App extends Component {
   constructor(props) {
@@ -21,6 +25,7 @@ class App extends Component {
 
   render() {
     return (
+      <React.StrictMode>
       <Router>
         <div className="App">
           <h1>Reactify</h1>
@@ -53,10 +58,10 @@ class App extends Component {
           </ul>
           <Route exact path="/" exact component={Home}/>
           <Route path="/login" component={Login}/>
-          <Route path="/albums" component={Albums}/>
+          <Route path="/albums" component={Loader(Albums)}/>
       </div>
     </Router>
-      
+    </React.StrictMode>
       
     );
   }
