+++ src/App/Services/Loader.js
@@ -1,0 +1,10 @@
+import React, { Suspense } from "react";
+
+export default function (Component) {
+    return props => (
+      <Suspense fallback={<div>Loading...</div>}>
+        <Component {...props} />
+      </Suspense>
+    );
+  }
+
